Update description of issue 6
I have now managed to configure a voicemail using the Voicemail Twimlet
from Twilio Labs:

  https://www.twilio.com/labs/twimlets/voicemail

This is a satisfying solution, although technically, it is not based on TwiML alone:
it relies on a hosted script in PHP which can only be customized through its
fixed set of parameters. Notably, it does not support the configuration of the
voice and language of the message, but it allows to play an audio recording
instead, which makes it useful beyond English-speaking countries.