Add script to list last missed calls
In our current setup, when a call is missed because our SIP client
is not connected, e.g. our computer is off or we did not launch
the SIP client today, we receive an error email from Twilio
(only once per day). Sadly, the email does not contain any more
details about the source of the error.

This script allows to retrieve the number of the last caller
after receiving such an error notification from Twilio by email.