Add script to retrieve latest SMS messages
Following the announcement on November 19th of the upcoming acquisition
of Pipedream by Workday, this script provides a backup plan in case
Pipedream workflows get discontinued or operate in a degraded manner
post-acquisition. Running this script allows to retrieve latest
SMS messages on demand.

Both sent and received messages are included, as the Messages API does
not allow filtering on the inbound/outbound 'direction' like the Calls API,
only on specific from/to phone numbers. I considered adding an extra
parameter to specify the phone numbers attached to my own Twilio account,
but I did not find the inconvenience of the missing filtering serious enough
to justify it.